+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.45.0] - 2026-06-11
+
+### Changed
+- `plot_upset()`, `plot_tax_upset()`: switch from ggupset to ggVennDiagram for upset plot generation
+
+### Removed
+- `plot_upset()`, `plot_tax_upset()`: removed `fill`, `show_counts`, `order_by` parameters (not applicable to ggVennDiagram); added `relative_height`, `relative_width` parameters
+
 ## [1.44.6] - 2026-06-11
 
 ### Changed